refactor(conversation): split acp error recovery
